requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pandas==2.1.4
numpy==1.26.3
python-multipart==0.0.6
pymongo==4.6.1
pymssql==2.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
msal==1.26.0
requests==2.31.0
openpyxl==3.1.2
python-dotenv
